Reorganize syncdb_helpers and add adjust columns
