+++ backend/src/xfd_django/xfd_api/management/commands/syncdb.py
@@ -2,9 +2,13 @@
 # Third-Party Libraries
 from django.core.management.base import BaseCommand
 from xfd_api.tasks.searchSync import handler as sync_es_domains
-from xfd_api.tasks.syncdb_helpers import (
-    drop_all_tables,
+
+# cisagov Libraries
+from backend.src.xfd_django.xfd_api.tasks.helpers.syncdb_helpers.create_sampe_data import (
     populate_sample_data,
+)
+from backend.src.xfd_django.xfd_api.tasks.syncdb_task import (
+    drop_all_tables,
     synchronize,
 )
