+++ backend/src/xfd_django/xfd_api/management/commands/syncmdl.py
@@ -5,12 +5,21 @@ import os
 # Third-Party Libraries
 from django.core.management.base import BaseCommand
 from django.db import connections
-from xfd_api.tasks.searchSync import handler as sync_es_domains
-from xfd_api.tasks.syncdb_helpers import (
-    drop_all_tables,
-    manage_elasticsearch_indices,
+
+# cisagov Libraries
+from backend.src.xfd_django.xfd_api.tasks.helpers.syncdb_helpers.create_sampe_data import (
     populate_sample_data,
+)
+from backend.src.xfd_django.xfd_api.tasks.helpers.syncdb_helpers.es_sync import (
+    manage_elasticsearch_indices,
     sync_es_organizations,
+)
+from backend.src.xfd_django.xfd_api.tasks.helpers.syndb_helpers.adjust_columns import (
+    adjust_column_types,
+)
+from backend.src.xfd_django.xfd_api.tasks.searchSync import handler as sync_es_domains
+from backend.src.xfd_django.xfd_api.tasks.syncdb_task import (
+    drop_all_tables,
     synchronize,
 )
 
@@ -107,6 +116,9 @@ class Command(BaseCommand):
             self.stdout.write("Dropping and recreating the database...")
             drop_all_tables(app_label="xfd_mini_dl")
         synchronize(target_app_label="xfd_mini_dl")
+
+        self.stdout.write("Running Phase 2 column type adjustments …")
+        adjust_column_types(target_app_label="xfd_mini_dl")
 
         self.stdout.write("Database synchronization complete.")
