+++ backend/src/xfd_django/xfd_api/tasks/refresh_vs_summaries.py
@@ -6,7 +6,6 @@ import os
 logging.basicConfig(level=logging.INFO, format="%(levelname)s: %(message)s")
 LOGGER = logging.getLogger(__name__)
 # Third-Party Libraries
-from xfd_api.tasks.syncdb_helpers import build_fake_host_summaries
 from xfd_api.tasks.vulnScanningSync import (
     create_daily_host_summary,
     create_port_scan_service_summaries,
@@ -15,6 +14,9 @@ from xfd_api.tasks.vulnScanningSync import (
     enforce_latest_flag_port_scan,
 )
 from xfd_mini_dl.models import Organization
+
+# cisagov Libraries
+from backend.src.xfd_django.xfd_api.tasks.syncdb_task import build_fake_host_summaries
 
 
 def rebuild_org_id_dict(db_name="mini_data_lake"):
