+++ backend/src/xfd_django/xfd_api/tests/test_domain.py
@@ -8,12 +8,6 @@ from django.db import connections, transaction
 from fastapi.testclient import TestClient
 import pytest
 from xfd_api.auth import create_jwt_token
-from xfd_api.tasks.syncdb_helpers import (
-    create_domain_view,
-    create_service_view,
-    create_vuln_materialized_views,
-    create_vuln_normal_views,
-)
 from xfd_django.asgi import app
 from xfd_mini_dl.models import (
     DataSource,
@@ -27,6 +21,13 @@ from xfd_mini_dl.models import (
     SubDomains,
     User,
     UserType,
+)
+
+# cisagov Libraries
+from backend.src.xfd_django.xfd_api.tasks.helpers.syncdb_helpers.create_db_views import (
+    create_domain_materialized_view,
+    create_service_mat_view,
+    create_vuln_materialized_views,
 )
 
 client = TestClient(app)
@@ -113,9 +114,8 @@ def sample_domain_ip_vuln(organization):
 def ensure_vuln_views_created(django_db_setup, django_db_blocker):
     """Ensure all necessary views for vulnerability testing are created."""
     with django_db_blocker.unblock():
-        create_domain_view("mini_data_lake")
-        create_vuln_normal_views("mini_data_lake")
-        create_service_view("mini_data_lake")
+        create_domain_materialized_view("mini_data_lake")
+        create_service_mat_view("mini_data_lake")
         create_vuln_materialized_views("mini_data_lake")
