+++ backend/src/xfd_django/xfd_api/tests/test_vulnerabilities.py
@@ -8,12 +8,6 @@ from django.db import transaction
 from fastapi.testclient import TestClient
 import pytest
 from xfd_api.auth import create_jwt_token
-from xfd_api.tasks.syncdb_helpers import (
-    create_domain_view,
-    create_service_view,
-    create_vuln_materialized_views,
-    create_vuln_normal_views,
-)
 from xfd_django.asgi import app
 from xfd_mini_dl.models import (
     DataSource,
@@ -30,6 +24,13 @@ from xfd_mini_dl.models import (
     UserType,
     Vulnerability,
     VulnScan,
+)
+
+# cisagov Libraries
+from backend.src.xfd_django.xfd_api.tasks.helpers.syncdb_helpers.create_db_views import (
+    create_domain_materialized_view,
+    create_service_mat_view,
+    create_vuln_materialized_views,
 )
 
 client = TestClient(app)
@@ -255,16 +256,14 @@ def ticket_vuln_setup(db):
 def ensure_vuln_views_created(django_db_setup, django_db_blocker):
     """Ensure all necessary views for vulnerability testing are created."""
     with django_db_blocker.unblock():
-        create_domain_view("mini_data_lake")
-        create_vuln_normal_views("mini_data_lake")
-        create_service_view("mini_data_lake")
+        create_domain_materialized_view("mini_data_lake")
+        create_service_mat_view("mini_data_lake")
 
 
 @pytest.fixture
 def refresh_vuln_views(django_db_blocker):
     """Refresh the materialized vuln views after data is inserted."""
     with django_db_blocker.unblock():
-        create_vuln_normal_views("mini_data_lake")
         create_vuln_materialized_views("mini_data_lake")
